third_party/blink/web_tests/platform/generic/fast/css/auto-min-size-expected.txt

PASS div.style.minWidth is ""
PASS getComputedStyle(div)["min-width"] is "0px"
PASS div.style.minWidth is "auto"
PASS div.style.maxWidth is ""
PASS getComputedStyle(div).minWidth is "0px"
PASS div.style.minWidth is "0px"
PASS getComputedStyle(div)["min-width"] is "0px"
PASS div.style.minHeight is ""
PASS getComputedStyle(div)["min-height"] is "0px"
PASS div.style.minHeight is "auto"
PASS div.style.maxHeight is ""
PASS getComputedStyle(div).minHeight is "0px"
PASS div.style.minHeight is "0px"
PASS getComputedStyle(div)["min-height"] is "0px"
PASS flexitem.style.minWidth is ""
PASS getComputedStyle(flexitem)["min-width"] is "auto"
PASS flexitem.style.minWidth is "auto"
PASS flexitem.style.maxWidth is ""
PASS getComputedStyle(flexitem).minWidth is "auto"
PASS flexitem.style.minWidth is "0px"
PASS getComputedStyle(flexitem)["min-width"] is "0px"
PASS flexitem.style.minHeight is ""
PASS getComputedStyle(flexitem)["min-height"] is "auto"
PASS flexitem.style.minHeight is "auto"
PASS flexitem.style.maxHeight is ""
PASS getComputedStyle(flexitem).minHeight is "auto"
PASS flexitem.style.minHeight is "0px"
PASS getComputedStyle(flexitem)["min-height"] is "0px"
PASS griditem.style.minWidth is ""
PASS getComputedStyle(griditem)["min-width"] is "auto"
PASS griditem.style.minWidth is "auto"
PASS griditem.style.maxWidth is ""
PASS getComputedStyle(griditem).minWidth is "auto"
PASS griditem.style.minWidth is "0px"
PASS getComputedStyle(griditem)["min-width"] is "0px"
PASS griditem.style.minHeight is ""
PASS getComputedStyle(griditem)["min-height"] is "auto"
PASS griditem.style.minHeight is "auto"
PASS griditem.style.maxHeight is ""
PASS getComputedStyle(griditem).minHeight is "auto"
PASS griditem.style.minHeight is "0px"
PASS getComputedStyle(griditem)["min-height"] is "0px"
PASS successfullyParsed is true

TEST COMPLETE
